refactor: centralize URL parsing utilities
